dashboard/requirements.txt
flask==2.1.3
werkzeug==2.1.2
psycopg2-binary==2.9.5